// frontend-v2/src/app/core/api/models/cliente-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClienteEnderecoOutput } from './cliente-endereco-output';

export interface ClienteOutput {
  id: number;
  nome: string;
  tel_1?: string;
  tel_2?: string;
  tel_3?: string;
  documento?: string;
  enderecos: Array<ClienteEnderecoOutput>;
}